refactor: resolve-dependencies
