+++ packages/resolve-dependencies/package.json
@@ -45,6 +45,8 @@
     "dependency-path": "workspace:8.0.6",
     "encode-registry": "^3.0.0",
     "get-npm-tarball-url": "^2.0.3",
+    "is-inner-link": "^4.0.0",
+    "is-subdir": "^1.1.1",
     "path-exists": "^4.0.0",
     "ramda": "^0.27.1",
     "replace-string": "^3.1.0",
